Removing Process_cart_txn from CP7
Due to Matt's investigation, an error in the DB copy script causes carts
not to be associated correctly with copied databases.  Until a fix can
be developed, we'll remove this from the test suite.  In addition, I'm
adding some code to close the erroring window so if something similar
happens in the future, it will only fail its own test, not cause
cascading fails.