Tempoary change to point FH at Prod
I'm trying to determine if it's feasible to merge over most of our
Fullhouse changes into our Production Automation branch.  The easiest
way to test this is to run the Fullhouse suite on Prod and go through
the results.  I will change the Fullhouse job back once testing is
completed.