Update for Query test in Bugs
Though this test consistently passes locally, it consistently fails in
AWS.  It appears that the 'Favorite Queries' name was causing the
failure, since it said it was already used.  I couldn't find anywhere
this was being created, nor was it present in the master database, but
the same failure continues to come up.  I'm changing the name of the
category, since it's only used to find the query created for this test.