Updating NCOA for USPS site changes
We were verifying we were on the correct USPS page by using a URL
present on the page.  USPS updated their site and updated the url to use
https rather than http, causing the check to fail.  I updated the test
to reflect this change and everything is working correctly.