Adding additional step to home page test
When verifying we're on the home page, a situation can develop where the
tiles don't load, showing 'Retrieving content' instead.  Usually this is
corrected by reloading the page, which the existing code does.  I'm
adding code to continue checking for Retrieving content each time the
page reloads.  If this results in a loop, I'll reconsider the fix.
Otherwise, this should continue refreshing the page, after a wait, to
load the page properly.  If this doesn't fix the situation, we may need
dev to look at why Retrieving content is getting stuck on the main page,
especially in automation.