Attempting to fix DIY issue in AWS
I'm trying to fix the DIY issue in AWS in a different way than before.
I thought this was just a timing issue, but it appears that something
else is happening in this case.  After completing the test, the test
will now fully navigate back to the DIY page, its final state.