Updating the way we navigate to KR
To facilitate the switch from FH to KR, I changed how we navigated to
'Fullhouse' which is the step used to navigate back to the main site
page.  Since we don't want to have to remember to change this back each
time we merge to prod, I'm changing it to simply update the site_url in
FH automation to point to KR inststead.  This is already updated when we
merge to prod so it won't result in any additional changes after a
merge.